requirements.txt
# Core libraries for numerical operations and image processing
numpy
pillow
matplotlib
opencv-python

# Libraries for AI model implementation and evaluation
scipy
scikit-learn
scikit-image

# Libraries for testing and documentation
pytest
sphinx
